+++ postgres/src/driver/tls.rs
@@ -43,7 +43,7 @@ pub(super) fn dangerous_config(alpn: Vec<Vec<u8>>) -> std::sync::Arc<xitca_tls::
                 message,
                 cert,
                 dss,
-                &rustls::crypto::ring::default_provider().signature_verification_algorithms,
+                &rustls::crypto::aws_lc_rs::default_provider().signature_verification_algorithms,
             )
         }
 
@@ -57,12 +57,12 @@ pub(super) fn dangerous_config(alpn: Vec<Vec<u8>>) -> std::sync::Arc<xitca_tls::
                 message,
                 cert,
                 dss,
-                &rustls::crypto::ring::default_provider().signature_verification_algorithms,
+                &rustls::crypto::aws_lc_rs::default_provider().signature_verification_algorithms,
             )
         }
 
         fn supported_verify_schemes(&self) -> Vec<rustls::SignatureScheme> {
-            rustls::crypto::ring::default_provider()
+            rustls::crypto::aws_lc_rs::default_provider()
                 .signature_verification_algorithms
                 .supported_schemes()
         }